+++ CarUPO/app/Http/Controllers/CochesController.php
@@ -198,7 +198,19 @@ class CochesController extends Controller
     }
     public function eliminarCoche(Request $request)
     {
+
         $coche = Coche::findOrFail($request->id);
+        $producto = $coche->producto;
+        if ($producto->lineas_de_compra->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociado a lineas de compra.');
+        }
+        if ($producto->lineas_de_carrito->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociado a lineas de carrito.');
+        }
+        if ($producto->favoritos_productos->count() > 0) {
+            return redirect()->back()->with('error', 'No se puede eliminar el coche porque está asociada a favoritos.');
+        }
+        $producto->delete();
         $coche->delete();
         return app()->make(ProductosController::class)->callAction('mostrarProductos', []);
     }
